generify and redundant casts
git-svn-id: http://twcore.org/svn/trunk@1534 9f8b3893-2a26-0410-9d00-a71e40fd2f77

+++ twcore/src/twcore/core/events/FrequencyChange.java
@@ -24,8 +24,8 @@ public class FrequencyChange extends SubspaceEvent {
 	 * @param array the ByteArray containing the packet data
 	 */
     public FrequencyChange(ByteArray array){
-        m_playerID = (short)array.readLittleEndianShort( 1 );
-        m_frequency = (short)array.readLittleEndianShort( 3 );
+        m_playerID = array.readLittleEndianShort( 1 );
+        m_frequency = array.readLittleEndianShort( 3 );
     }
 
     /**
